fixed overflow for home page and added Header as a component so that it can be imported by all other components
