+++ src/Components/Checkout/Checkout.style.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const Wrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  width: 100vw;
+  margin-top: 20px;
+  align-items: center;
+`;
