+++ src/Components/Home/Home.style.js
@@ -8,30 +8,6 @@ export const Wrapper = styled.div`
   align-items: center;
 `;
 
-export const Header = styled.div`
-  position: absolute;
-  top: 1vh;
-  left: 2vw;
-  display: flex;
-  width: 96vw;
-  height: 16vh;
-  justify-content: space-between;
-  align-items: center;
-  background-color: #f3f3f3;
-  border-radius: 10px;
-  box-shadow: rgba(9, 30, 66, 0.25) 0px 4px 8px -2px,
-    rgba(9, 30, 66, 0.08) 0px 0px 0px 1px;
-
-  img {
-    height: 14vh;
-  }
-
-  .svg {
-    height: 5vh;
-    padding-right: 15px;
-  }
-`;
-
 export const ContentWrapper = styled.div`
   height: 80vh;
   width: 96vw;
@@ -43,7 +19,11 @@ export const ContentWrapper = styled.div`
   border-radius: 10px;
   box-shadow: rgba(9, 30, 66, 0.25) 0px 4px 8px -2px,
     rgba(9, 30, 66, 0.08) 0px 0px 0px 1px;
-  overflow-y: scroll;
+  overflow: hidden;
+
+  &:hover {
+    overflow-y: scroll;
+  }
 `;
 
 export const ProductWrapper = styled.div`
